+++ ibrdtn/daemon/src/storage/MemoryBundleStorage.h
@@ -94,7 +94,7 @@ namespace dtn
 			/**
 			 * @sa BundleStorage::count()
 			 */
-			unsigned int count();
+			size_t count();
 
 			/**
 			 * @sa BundleStorage::releaseCustody();
